requirements.txt
streamlit==1.45.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.0
openai==1.50.0
pinecone-client==3.2.2
pypdf==4.3.1
python-dotenv==1.0.1
tiktoken==0.7.0